# blockchain-api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: KALDRIX Blockchain API
  description: REST and WebSocket API for KALDRIX blockchain operations including DAG management, quantum-resistant cryptography, and PBFT consensus
  version: 1.0.0
  contact:
    name: KALDRIX Team
    email: api@kaldrix.io
    url: https://kaldrix.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.kaldrix.io
    description: Production server
paths:
...
  /api/v1/blockchain/dag:
    get:
      summary: Get DAG State
      description: Retrieve current DAG state with nodes and metrics
      tags:
        - Blockchain
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: DAG state retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Block'
                  total:
                    type: integer
                  tips:
                    type: array
                    items:
                      type: string
                  metrics:
                    type: object
  /api/v1/blockchain/dag/nodes/{nodeId}:
    get:
      summary: Get Specific Node
      description: Retrieve a specific DAG node by ID
      tags:
        - Blockchain
      security:
        - bearerAuth: []
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Node retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
        '404':
          description: Node not found
  /api/v1/blockchain/dag/tips:
    get:
      summary: Get DAG Tips
      description: Retrieve current DAG tips (unconfirmed transactions)
      tags:
        - Blockchain
      security:
        - bearerAuth: []
      parameters:
        - name: max
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Tips retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tips:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  count:
                    type: integer
  /api/v1/blockchain/status:
    get:
      summary: Get Blockchain Status
      description: Retrieve blockchain status and metrics
      tags:
        - Blockchain
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  metrics:
                    type: object
                  version:
                    type: string

  # Transaction Operations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Transaction:
      type: object
      required:
        - id
        - sender
        - receiver
        - amount
        - timestamp
        - signature
      properties:
        id:
          type: string
          example: "tx_0x123abc456def789ghi012jkl345mno678pqr901stu"
        sender:
          type: string
          format: address
          example: "kx1q9f5j8g7h6d5s4a3z2x1c9v8b7n6m5k4j3i2h1g"
        receiver:
          type: string
          format: address
          example: "kx1a2b3c4d5e6f7g8h9i0j1k2l3m4n5o6p7q8r9s0t"
        amount:
          type: number
          example: 100.5
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        signature:
          type: string
          example: "0x123abc456def789ghi012jkl345mno678pqr901stu..."
        parents:
          type: array
          items:
            type: string
          description: Parent transaction IDs in DAG
        status:
          type: string
          enum: [pending, confirmed, failed]
          example: "confirmed"
        fee:
          type: number
          example: 0.001
        data:
          type: object
          description: Additional transaction data
    Block:
      type: object
      required:
        - id
        - transactions
        - timestamp
        - validator
        - signature
      properties:
        id:
          type: string
          example: "block_0x123abc456def789ghi012jkl345mno678pqr901stu"
        transactions:
          type: array
          items:
            type: string
          description: Array of transaction IDs
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        validator:
          type: string
          format: address
          example: "kx1v1l2i3d4a5t6o7r8a9d0d1r2e3s4s5i6g7n8a9t0u"
        signature:
          type: string
          example: "0x123abc456def789ghi012jkl345mno678pqr901stu..."
        parents:
          type: array
          items:
            type: string
          description: Parent block IDs in DAG
        height:
          type: integer
          example: 12345
        hash:
          type: string
          example: "0x123abc456def789ghi012jkl345mno678pqr901stu"
        merkleRoot:
          type: string
          example: "0x123abc456def789ghi012jkl345mno678pqr901stu"
        nonce:
          type: integer
          example: 42
...
security:
  - bearerAuth: []
  - apiKeyAuth: []